requirements.txt
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=0.24.2,<1.0.0
tensorflow-macos>=2.6.0
colorama>=0.4.4
python-dateutil>=2.8.2
pyfiglet>=0.8.post1
rich>=10.12.0
watchdog>=2.1.6
elasticsearch>=7.14.0
requests>=2.26.0
python-dotenv>=0.19.0
PyYAML>=5.4.1 